refacotring towards final str
