+++ src/main/scala/co/topl/quivr/archive/TestAlgebra.worksheet.sc
