+++ src/main/scala/co/topl/quivr/archive/algebras/DigestAlgebra.scala
