+++ src/main/scala/co/topl/quivr/archive/algebras/QuivrAlgebra.scala
