+++ src/main/scala/co/topl/quivr/archive/algebras/QuivrContract.scala
