+++ src/main/scala/co/topl/quivr/archive/algebras/QuivrContractProposer.scala
