+++ src/main/scala/co/topl/quivr/archive/algebras/QuivrContractProver.scala
