+++ src/main/scala/co/topl/quivr/archive/algebras/QuivrContractVerifier.scala
