+++ src/main/scala/co/topl/quivr/archive/algebras/SignerAlgebra.scala
