+++ src/main/scala/co/topl/quivr/archive/algebras/ValidateAlgebra.scala
