+++ src/main/scala/co/topl/quivr/archive/algebras/package.scala
