+++ src/main/scala/co/topl/quivr/archive/interpreters/NativeIoTransaction.scala
