+++ src/main/scala/co/topl/quivr/archive/interpreters/Proofs.scala
