+++ src/main/scala/co/topl/quivr/archive/interpreters/Propositions.scala
