+++ src/main/scala/co/topl/quivr/archive/interpreters/SpentTransactionOutput.scala
