+++ src/main/scala/co/topl/quivr/archive/interpreters/Verifiers.scala
