+++ src/main/scala/co/topl/quivr/archive/v1/Proof.scala
@@ -1,0 +1,8 @@
+package co.topl.quivr.v1
+
+
+sealed abstract class Proof
+
+object Proofs {
+    case class Digest(preimage: Array[Byte]) extends Proof
+}
