+++ src/main/scala/co/topl/quivr/archive/v1/Proposition.scala
@@ -1,0 +1,7 @@
+package co.topl.quivr.v1
+
+sealed abstract class Proposition 
+
+object Propositions {
+    case class Digest(digest: Array[Byte]) extends Proposition
+}
