+++ src/main/scala/co/topl/quivr/archive/v1/Verification.scala
@@ -1,0 +1,17 @@
+package co.topl.quivr.v1
+
+import cats._
+import cats.data.OptionT
+import cats.implicits._
+import co.topl.crypto.hash.blake2b256
+
+object Verification {
+
+  def digest[F[_]: Applicative](
+      proposition: Propositions.Digest,
+      proof: Proofs.Digest
+  ): F[Boolean] =
+    (blake2b256.hash(proof.preimage).value sameElements proposition.digest)
+      .pure[F]
+
+}
