+++ src/main/scala/co/topl/quivr/archive/v2/Proof.scala
@@ -1,0 +1,8 @@
+package co.topl.quivr.v2
+
+
+sealed abstract class Proof
+
+object Proofs {
+    case class Digest(preimage: Array[Byte], witness: Array[Byte]) extends Proof
+}
