+++ src/main/scala/co/topl/quivr/archive/v2/Proposition.scala
@@ -1,0 +1,11 @@
+package co.topl.quivr.v2
+
+sealed abstract class Proposition
+
+object Propositions {
+  case class Digest(digest: Array[Byte]) extends Proposition
+
+  object Digest {
+    val token: Byte = 1: Byte
+  }
+}
